Improved data visualization.
If the categorical data have more than 5 unique values it will visualize the bar chart horizontally, and if it has less than 5 it will visualize the bar chart vertically.

There is also a set amount of unique values that's allowed, so columns like "Date" with so many time values will not be visualize.

If there is only one category value in the column it will just display the category and count.